requirements.txt
line-bot-sdk
bs4
flask
datetime
emoji
pyshorteners
gunicorn
pymongo
dnspython